+++ src/fonduer/candidates/models/document_mention.py
@@ -6,6 +6,7 @@ from sqlalchemy.orm import relationship
 from fonduer.candidates.models.temporary_context import TemporaryContext
 from fonduer.parser.models import Document
 from fonduer.parser.models.context import Context
+from fonduer.parser.models.utils import construct_stable_id
 
 
 class TemporaryDocumentMention(TemporaryContext):
@@ -43,8 +44,14 @@ class TemporaryDocumentMention(TemporaryContext):
         return hash(self.document)
 
     def get_stable_id(self) -> str:
-        """Return a stable id for the ``DocumentMention``."""
-        return self.document.stable_id
+        """
+        Return a stable id.
+
+        :rtype: string
+        """
+        return construct_stable_id(
+            self.document, self._get_polymorphic_identity(), 0, 0
+        )
 
     def _get_table(self) -> Type["DocumentMention"]:
         return DocumentMention
@@ -82,3 +89,7 @@ class DocumentMention(Context, TemporaryDocumentMention):
         "polymorphic_identity": "document_mention",
         "inherit_condition": (id == Context.id),
     }
+
+    def __init__(self, tc: TemporaryDocumentMention):
+        self.stable_id = tc.get_stable_id()
+        self.document = tc.document