+++ src/fonduer/parser/models/utils.py
@@ -1,4 +1,4 @@
-from typing import Tuple
+from typing import List, Tuple
 
 from fonduer.parser.models import Context
 
@@ -15,13 +15,30 @@ def construct_stable_id(
     Contruct a stable ID for a Context given its parent and its character
     offsets relative to the parent.
     """
-    doc_id, _, parent_doc_char_start, _ = split_stable_id(parent_context.stable_id)
+
+    doc_id, _, idx = split_stable_id(parent_context.stable_id)
+
+    # Caption, document, figure, paragraph, section, table
+    if len(idx) == 1:
+        parent_doc_start = idx[0]
+        return f"{doc_id}::{polymorphic_type}:{parent_doc_start}"
+    # Cell
+    elif len(idx) == 3:
+        cell_pos = idx[0]
+        cell_row_start = idx[1]
+        cell_col_start = idx[2]
+        return (
+            f"{doc_id}::{polymorphic_type}:{cell_pos}:{cell_row_start}:{cell_col_start}"
+        )
+
+    # Span
+    parent_doc_char_start = idx[0]
     start = parent_doc_char_start + relative_char_offset_start
     end = parent_doc_char_start + relative_char_offset_end
     return f"{doc_id}::{polymorphic_type}:{start}:{end}"
 
 
-def split_stable_id(stable_id: str) -> Tuple[str, str, int, int]:
+def split_stable_id(stable_id: str,) -> Tuple[str, str, List[int]]:
     """Split stable id, returning:
 
         * Document (root) stable ID
@@ -33,6 +50,8 @@ def split_stable_id(stable_id: str) -> Tuple[str, str, int, int]:
     split1 = stable_id.split("::")
     if len(split1) == 2:
         split2 = split1[1].split(":")
-        if len(split2) == 3:
-            return split1[0], split2[0], int(split2[1]), int(split2[2])
+        type = split2[0]
+        idx = [int(_) for _ in split2[1:]]
+        return split1[0], type, idx
+
     raise ValueError(f"Malformed stable_id:\t{stable_id}")